backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-dateutil
python-jose
passlib
python-multipart
psycopg2-binary
python-dotenv
alembic